Move outline extraction into the pdf widget
